+++ NodeBindings/CHANGELOG.md
@@ -1,5 +1,6 @@
 [25/04/2017]
 - Retry TakePhoto when it fails with CaptureError state change
+- Return path to downloaded photo or video
 
 [24/04/2017]
 - 1300D TakePhoto works properly when live view is started before